Move worker thread function to separate file
